fix: Move ProjectCard to client component to fix 500 error
The profile page was returning HTTP 500 because it contained server
components with onClick handlers (window.open) and onError handlers
that require client-side JavaScript.

Changes:
- Extract ProjectCard and ProjectScreenshot to a separate client
  component file with "use client" directive
- Import ProjectCard in the main page.tsx (server component)
- Remove unused imports (Star, ExternalLink, Code, MentionsRow)
- Remove duplicate constant definitions (LANGUAGE_COLORS, etc)

This separation allows the main page to render server-side while
the interactive ProjectCard components hydrate on the client.
